+++ trackstream/utils/coordinates.py
@@ -40,7 +40,10 @@ DataType = T.Union[CoordinateType, CoordinateRepresentationType]
 
 
 def cartesian_to_spherical(
-    x: T.Sequence, y: T.Sequence, z: T.Sequence, deg: bool = False
+    x: T.Sequence,
+    y: T.Sequence,
+    z: T.Sequence,
+    deg: bool = False,
 ) -> T.Tuple[T.Sequence, T.Sequence, T.Sequence]:
     """Cartesian to Spherical.